+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # v2.0.0 (unreleased)
 
+*   Log transforms. You can now supply transforms to modify the buffered
+    contents before flushing them to the log device.
+
 *   'buffering_logger/rails' has been deprecated.  Instead, require
     'buffering_logger/railtie' and call 'BufferingLogger::Railtie.install'.